backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
openai==0.27.8
python-dotenv==1.0.0
websockets==11.0.3
idna==3.10